+++ app/kuma-cp/cmd/cmd_standalone_postgres_test.go
@@ -3,10 +3,25 @@
 package cmd
 
 import (
+	"github.com/Kong/kuma/pkg/config"
+	kuma_cp "github.com/Kong/kuma/pkg/config/app/kuma-cp"
+	"github.com/Kong/kuma/pkg/config/core/resources/store"
+
 	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
 )
 
 var _ = Describe("Standalone Postgres test", func() {
+
+	BeforeEach(func() {
+		// setup migrate DB
+		cfg := kuma_cp.DefaultConfig()
+		err := config.Load("", &cfg)
+		cfg.Store.Type = store.PostgresStore
+		Expect(err).ToNot(HaveOccurred())
+		err = migrate(cfg)
+		Expect(err).ToNot(HaveOccurred())
+	})
 
 	RunSmokeTest(StaticConfig(`
 xdsServer: